requirements.txt
streamlit 
google-generativeai 
python-dotenv 
pdf2image 
fpdf